+++ app/view/link/link.php
@@ -7,21 +7,27 @@
         <th class="text-center"><i class="far fa-edit"></i></th>
         <th class="text-center"><i class="fas fa-trash-alt"></i></th>
     </thead>
-    <?php foreach($links as $link): ?>
-        <tr>
-            <td class="text-center"><?php echo $link->getId(); ?></td>
-            <td class="text-center"><?php echo $link->getName(); ?></td>
-            <td class="text-center"><?php echo $link->getLink(); ?></td>
-            <?php foreach($this->articleModel->getAll() as $article): ?>
-                <?php if ($article->getId() == $link->getIdArticle()) echo "<td class='text-center'>".$article->getName()."</td>"; ?>
-            <?php endforeach; ?>
+    <?php if ($links != null) { ?>
+        <?php foreach($links as $link): ?>
+            <tr>
+                <td class="text-center"><?php echo $link->getId(); ?></td>
+                <td class="text-center"><?php echo $link->getName(); ?></td>
+                <td class="text-center"><?php echo $link->getLink(); ?></td>
+                <?php foreach($this->articleModel->getAll() as $article): ?>
+                    <?php if ($article->getId() == $link->getIdArticle()) echo "<td class='text-center'>".$article->getName()."</td>"; ?>
+                <?php endforeach; ?>
 
-            <td class="text-center">
-                <a class="btn btn-warning" href="?c=link&a=edit&id=<?php echo $link->getId(); ?>">Edit</a>
-            </td>
-            <td class="text-center">
-                <a class="btn btn-danger" onclick="javascript:return confirm('Do you want delete this link?');" href="?c=link&a=delete&id=<?php echo $link->getId(); ?>">Delete</a>
-            </td>
+                <td class="text-center">
+                    <a class="btn btn-warning" href="?c=link&a=edit&id=<?php echo $link->getId(); ?>">Edit</a>
+                </td>
+                <td class="text-center">
+                    <a class="btn btn-danger" onclick="javascript:return confirm('Do you want delete this link?');" href="?c=link&a=delete&id=<?php echo $link->getId(); ?>">Delete</a>
+                </td>
+            </tr>
+        <?php endforeach; ?>
+    <?php } else { ?>
+        <tr>
+            <td colspan="6" class="text-center text-info bg-light"> THERE ARE NO RECORDS </td>
         </tr>
-    <?php endforeach; ?>
+    <?php } ?>
 </table>
